Created InputData class for readability.
